Fix multi-marker case
htmltools functions are not vectorised. In addition, attributes from
htmltools::HTML are required by the markers tab to format links
properly but are lost when unlist-ing. There doesn't seem to be a way
to map to an atomic vector. So unlist and then force some attributes
into place afterwards - structure is vectorised.